% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CosGeneGateDeconvRBMode.R
\name{Calculate_RMSE}
\alias{Calculate_RMSE}
\title{This is the function for calculating root mean square error}
\usage{
Calculate_RMSE(decon, gt)
}
\arguments{
\item{decon}{File directory for single cell reference file.}

\item{gt}{File directory for realbulk reference.}
}
\value{
Return the RMSE value
}
\description{
This is the function for calculating root mean square error
}
