% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CosGeneGateDeconvPBMode.R
\name{create_mixture}
\alias{create_mixture}
\title{This is to create a pseudo-bulk mixture}
\usage{
create_mixture(df_demo_test, ct_list)
}
\arguments{
\item{df_demo_test}{The single cell profile used to generate pseudo-bulk mixtures.}

\item{ct_list}{List of existing cell types.}
}
\value{
Return one pseudo-bulk mixture and its ground truth.
}
\description{
This is to create a pseudo-bulk mixture
}
