% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CosGeneGateDeconvPBMode.R
\name{create_mixture_helper}
\alias{create_mixture_helper}
\title{This is the helper function for generating pseudo-bulk mixture}
\usage{
create_mixture_helper(ct_chosen, lower_thres = 0.05)
}
\arguments{
\item{ct_chosen}{Selected cell types in the mixture.}

\item{lower_thres}{The minimum proportion of a cell type.}
}
\value{
Return the cell type proportion for selected cell types.
}
\description{
This is the helper function for generating pseudo-bulk mixture
}
