% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CosGeneGateDeconvRBMode.R
\name{deconvolution_realbulk_mode}
\alias{deconvolution_realbulk_mode}
\title{This is the function for CosGeneGate deconvolution - realbulk mode}
\usage{
deconvolution_realbulk_mode(
  sc_profile,
  realbulk_profile,
  realbulk_gt,
  mkg_file,
  index = c(2, 3, 4, 5, 6, 7, 8, 9, 10, 12, 14, 16, 18, 20, 25, 30, 35, 40, 45, 50),
  plot_line = FALSE
)
}
\arguments{
\item{sc_profile}{File directory for single cell reference file.}

\item{realbulk_profile}{File directory for realbulk reference.}

\item{realbulk_gt}{File directory for the groundtruth of realbulk reference.}

\item{mkg_file}{Marker gene file generated by CosGeneGate.}

\item{index}{Numbe of marker genes to be iterated.}

\item{plot_line}{TRUE to plot the graph of error vs. # of marker genes, FALSE otherwise.}
}
\value{
Return the optimized single cell reference profile.
}
\description{
This is the function for CosGeneGate deconvolution - realbulk mode
}
