% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CosGeneGateDeconvPBMode.R
\name{generate_pseudobulk}
\alias{generate_pseudobulk}
\title{This is to create n pseudo bulk mixtures and their ground truths}
\usage{
generate_pseudobulk(n, df, prop = 0.1)
}
\arguments{
\item{n}{Number of pseudo-bulk mixtures to be generated.}

\item{df}{Single cell data.}

\item{prop}{Proportion of cells used to generate pseudobulk mixtures.}
}
\value{
Return n pseudo-bulk mixture and their ground truth.
}
\description{
This is to create n pseudo bulk mixtures and their ground truths
}
